requirements.txt
numpy
Pillow
seaborn
torchvision